refactor(storage): move local/remote_s3/remote_gcs interfaces to backends/ sub-module
