+++ packages/devqubit-engine/src/devqubit_engine/storage/factory.py
@@ -40,12 +40,12 @@ from typing import Any, Protocol
 from urllib.parse import parse_qs, urlparse
 
 from devqubit_engine.core.config import Config, get_config
-from devqubit_engine.storage.errors import StorageError
-from devqubit_engine.storage.local import (
+from devqubit_engine.storage.backends.local import (
     LocalRegistry,
     LocalStore,
     LocalWorkspace,
 )
+from devqubit_engine.storage.errors import StorageError
 from devqubit_engine.storage.types import ObjectStoreProtocol, RegistryProtocol
 
 
@@ -88,7 +88,7 @@ def _lazy_import_s3() -> tuple[ObjectStoreBackend, RegistryBackend]:
     ImportError
         If devqubit-engine[s3] is not installed.
     """
-    from devqubit_engine.storage.remote_s3 import S3Registry, S3Store
+    from devqubit_engine.storage.backends.remote_s3 import S3Registry, S3Store
 
     return S3Store, S3Registry
 
@@ -107,7 +107,7 @@ def _lazy_import_gcs() -> tuple[ObjectStoreBackend, RegistryBackend]:
     ImportError
         If devqubit-engine[gcs] is not installed.
     """
-    from devqubit_engine.storage.remote_gcs import GCSRegistry, GCSStore
+    from devqubit_engine.storage.backends.remote_gcs import GCSRegistry, GCSStore
 
     return GCSStore, GCSRegistry
