+++ tests/conftest.py
@@ -20,7 +20,7 @@ from click.testing import CliRunner
 from devqubit_engine.cli import cli
 from devqubit_engine.core.config import Config
 from devqubit_engine.core.record import RunRecord
-from devqubit_engine.storage.local import LocalRegistry, LocalStore
+from devqubit_engine.storage.backends.local import LocalRegistry, LocalStore
 from devqubit_engine.storage.types import ArtifactRef
 from devqubit_engine.utils.common import utc_now_iso
